chore(web-components): refactor test setup and test/sb utils approach
