--- packages/web-components/src/__test__/setup-browser.cts
@@ -1,6 +1,0 @@
-function importAll(r: __WebpackModuleApi.RequireContext): void {
-  r.keys().forEach(r);
-}
-
-// Explicitly add to browser test
-importAll(require.context('../', true, /\.spec\.js$/));
