+++ packages/web-components/src/text/text.stories.ts
@@ -1,6 +1,6 @@
 import { html } from '@microsoft/fast-element';
 import type { Args, Meta } from '@storybook/html';
-import { renderComponent } from '../__test__/helpers.js';
+import { renderComponent } from '../helpers.js';
 import type { Text as FluentText } from './text.js';
 import './define.js';
 import { TextAlign, TextFont, TextSize, TextWeight } from './text.options.js';
